% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WorkflowEvaluationFunctions.R
\name{assign.status}
\alias{assign.status}
\title{assign.status}
\usage{
assign.status(x, status, data.type, version)
}
\arguments{
\item{Dataframe}{which represent sthe Evaluation dataset or the Reference Dataset}

\item{tag}{to determine the evaluation or reference status}

\item{Workflow}{category}

\item{Workflow}{path number}
}
\value{
the taged identifier for further processing
}
\description{
places tag on to an identifier
}

