% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WorkflowEvaluationFunctions.R
\name{expectedUtility}
\alias{expectedUtility}
\title{expectedUtility}
\usage{
expectedUtility(dataset, label = "", Lfp = 1, Utp = 1, deltaPlus = 1,
  guarantee = 1e-05)
}
\arguments{
\item{dataset}{the resulting dataframe from Workflow.posteriorestimate}

\item{a}{label for the workflow}

\item{Loss}{of a false positive}

\item{Utility}{of a true positive}
}
\value{
the expected utilty and associated estimates
}
\description{
Given the Posterior estimates evaluate the expected utility for a users determined Utp,Lfp,deltaPlus,guarantee
}

