% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WorkflowEvaluationFunctions.R
\name{fit2clusters.workflow}
\alias{fit2clusters.workflow}
\title{fit2clusters.workflow}
\usage{
fit2clusters.workflow(Y, Ysigsq, bootModel, piStart = c(0.5, 0.5),
  VStart = c(0.1, 0.1), psiStart = c(0, 0.1), NinnerLoop = 1,
  nReps = 500, psi0Constraint, V0Constraint, sameV = FALSE,
  estimatesOnly = TRUE, printMe = TRUE, plotMe = TRUE, testMe = FALSE,
  Ntest = 5000, seed)
}
\arguments{
\item{Multiple}{parameters - to be completed later}
}
\value{
"a dataframe of posterior odds and posterior probability variance"
}
\description{
"Performs the EM algorithm over the bootstrap values"
}

