% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WorkflowEvaluationFunctions.R
\name{make.workflow.map}
\alias{make.workflow.map}
\title{make.workflow.map}
\usage{
make.workflow.map(Merged.options)
}
\arguments{
\item{Merged.options}{A concatenation of the evaluation and reference data set features.}
}
\value{
A data frame with two columns: "drivers" (the IDs for the features in the reference data set), and "workflow_options_merged":
}
\description{
Make a workflow map
}
\details{
The "drivers" are the ID
}

