% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WorkflowEvaluationFunctions.R
\name{merge.tag.options}
\alias{merge.tag.options}
\title{Merge tag options}
\usage{
\method{merge}{tag.options}(Workflow.Data, ReferenceTag = "P",
  EvaluationTag = "RS")
}
\arguments{
\item{Workflow.Data}{object where reference is the first item an the evaluation data is all further items in list}

\item{ReferenceTag}{which sets a particular symbol for the reference set tag}

\item{EvaluationTag}{which sets a particular symbol for the evaluation set tag}
}
\value{
Merged.options a dataframe with renamed labels
}
\description{
Merge tag options
}

