% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WorkflowEvaluationFunctions.R
\name{PlotEvaluationTable}
\alias{PlotEvaluationTable}
\title{PlotEvaluationTable}
\usage{
PlotEvaluationTable(Evaluation.table, EvaluationOrder = "TEU")
}
\arguments{
\item{Evaluation}{Table}
}
\value{
Plot of Expected utility and number of filters applied
}
\description{
PlotEvaluationTable:  Produces a plot for the expected utility of workflow path
}

