% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WorkflowEvaluationFunctions.R
\name{Workflow.Criterion}
\alias{Workflow.Criterion}
\title{Workflow.Criterion}
\usage{
Workflow.Criterion(Model.quality.object)
}
\arguments{
\item{Model.quality.object}{the pairs with data to be used to estimate model quality}
}
\value{
"Values of model quality per pair"
}
\description{
User specifies the model quality used to evaluate the workflow
}

