% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WorkflowEvaluationFunctions.R
\name{Workflow.Evaluation.table}
\alias{Workflow.Evaluation.table}
\title{Workflow.Evaluation.table}
\usage{
Workflow.Evaluation.table(Posterior.dataframe, Lfp = 1, Utp = 1,
  deltaPlus = 1, guarantee = 1e-05)
}
\arguments{
\item{Posterior.dataframe}{Produced by Workflow.posteriorestimate().}
}
\value{
Nicely formatted table of posterior probabilities, Pr(+) and Pr(-), standard deviations, model quality scores, and biases.
}
\description{
Workflow.Evaluation.table:  Produces an expected utility table for guidance and evaluation
}

