% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WorkflowEvaluationFunctions.R
\name{Workflow.posteriorestimate}
\alias{Workflow.posteriorestimate}
\title{Workflow.posteriorestimate}
\usage{
Workflow.posteriorestimate(Model.quality.object, Model.Quality,
  postProb = NULL, postProbVar = NULL)
}
\arguments{
\item{Model}{quality data and Model Quality}
}
\value{
a evaulatioin ready dataframe of the posterior estimates
}
\description{
Calls the fit2clusters.workflow function and creates a dataframe with bootstap values, posterior odds, posterior probability variance, for each identifier pair
}

