% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WorkflowEvaluationFunctions.R
\name{WorkflowEvaluationData}
\alias{WorkflowEvaluationData}
\title{WorkflowEvaluationData}
\usage{
WorkflowEvaluationData(EvaluationExperimentSet, ReferenceSet)
}
\arguments{
\item{EvaluationExperimentSet}{Merged set of Workflow options on the same samples.}

\item{ReferenceSet}{The Reference data set which is related to the Evaluation set as determined by the Model quality}
}
\value{
An object of the class list which the first item is the reference dataset and all following list items are workflow options on the same samples
}
\description{
WorkflowEvaluationData
}

