% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WorkflowEvaluationFunctions.R
\name{assign.status}
\alias{assign.status}
\title{assign.status}
\usage{
assign.status(x, status, data.type, version)
}
\arguments{
\item{x}{A set of identifiers  TODO: is it really a "set"?}

\item{status}{TODO}

\item{data.type}{TODO}

\item{version}{TODO}
}
\value{
the taged identifier
}
\description{
places tag on to an identifier
}

