% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WorkflowEvaluationFunctions.R
\name{assign.status}
\alias{assign.status}
\title{assign.status}
\usage{
assign.status(x, status, data.type, version)
}
\arguments{
\item{a}{set of identifiers}
}
\value{
the taged identifier
}
\description{
places tag on to an identifier
}

