% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WorkflowEvaluationFunctions.R
\name{make.workflow.map}
\alias{make.workflow.map}
\title{make.workflow.map}
\usage{
make.workflow.map(Merged.options)
}
\arguments{
\item{Merged}{options}
}
\value{
WorkflowMap.object a dataframe of reference tags mapped to the respective evaluation tags
}
\description{
Make a workflow map
}

